# packages/worker/openapi.yaml
openapi: 3.1.0
info:
  title: AgentLair API
  version: "0.18.0"
  description: |
    AgentLair is an infrastructure layer for AI agents — email inboxes, encrypted secret storage,
    agent calendar, shared observations, and cross-agent coordination. All at your own domain.

    ## Authentication
    Most endpoints require an API key passed as a Bearer token:
    ```
    Authorization: Bearer al_live_<your_key>
    ```
    Session tokens (from magic link login) are also accepted: `Authorization: Bearer session_<token>`

    ## Rate Limits
    Free tier: 100 API requests/day, 10 emails/day. Paid tier: 10,000/day.
    When email limits are exceeded, the API returns HTTP 402 with x402 payment requirements
    (0.01 USDC on Base via the x402 protocol).

  contact:
    url: https://agentlair.dev
  license:
    name: Proprietary
servers:
  - url: https://agentlair.dev
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: al_live_<key> or session_<token>
...
security:
  - bearerAuth: []
paths:
...
  /v1/budget:
    get:
      tags: [budget]
      summary: Get spending caps and budget status
      description: |
        Returns per-period (daily/weekly/monthly) spending caps, current spend, and remaining budget.
        Also returns budget settings: `on_limit` behavior and `single_tx_limit`.

        All amounts are in atomic USDC units (1,000,000 = 1.00 USDC).
      responses:
        "200":
          description: Current budget state
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id:
                    type: string
                  currency:
                    type: string
                    example: USDC
                  atomic_unit:
                    type: string
                    example: "1e-6 USDC (one millionth of one USDC)"
                  caps:
                    type: object
                    description: Per-period spending caps. Null if no cap set for that period.
                    properties:
                      daily:
                        nullable: true
                        type: object
                        properties:
                          limit_usdc:
                            type: integer
                          spent_usdc:
                            type: integer
                          remaining_usdc:
                            type: integer
                          resets_at:
                            type: string
                            format: date-time
                      weekly:
                        nullable: true
                        type: object
                      monthly:
                        nullable: true
                        type: object
                  has_caps:
                    type: boolean
                  status:
                    type: string
                    enum: [active, limit_reached]
                  settings:
                    type: object
                    properties:
                      on_limit:
                        type: string
                        enum: [reject, approve]
                        description: What to do when budget is exceeded (reject = 402, approve = 202 + approval flow)
                      single_tx_limit_cents:
                        type: integer
                        nullable: true
                        description: Max allowed per-transaction amount (atomic USDC units). Null = no limit.
    put:
      tags: [budget]
      summary: Set spending caps
      description: |
        Configure per-period spending caps and behavior when limits are reached.

        **Flat format (recommended):** Send daily/weekly/monthly at top level without a `caps` wrapper.

        **on_limit:** Controls what happens when a charge exceeds the cap:
        - `reject` (default): Returns HTTP 402 — agent must stop or pay via x402.
        - `approve`: Returns HTTP 202 — creates an approval request for the principal to review.

        **single_tx_limit_cents:** Maximum per-transaction amount in atomic USDC units. Charges above this require approval even if within budget.

        All amounts in atomic USDC units (1,000,000 = 1.00 USDC).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                daily:
                  type: integer
                  nullable: true
                  description: Daily cap in atomic USDC units. Null to remove cap. Example 10000000 = 10 USDC/day.
                weekly:
                  type: integer
                  nullable: true
                  description: Weekly cap. Must be >= daily cap if both set.
                monthly:
                  type: integer
                  nullable: true
                  description: Monthly cap. Must be >= weekly cap if both set.
                caps:
                  type: object
                  description: Alternative nested format. Can use flat format (daily/weekly/monthly) instead.
                  properties:
                    daily:
                      type: integer
                      nullable: true
                    weekly:
                      type: integer
                      nullable: true
                    monthly:
                      type: integer
                      nullable: true
                on_limit:
                  type: string
                  enum: [reject, approve]
                  description: "What to do when budget exceeded. Default: reject"
                single_tx_limit_cents:
                  type: integer
                  nullable: true
                  description: Per-transaction limit in atomic USDC units. Null to remove.
            examples:
              flat_with_approval:
                summary: Set daily cap with approval flow
                value:
                  daily: 10000000
                  on_limit: approve
              nested_format:
                summary: Set caps using nested format
                value:
                  caps:
                    daily: 10000000
                    monthly: 50000000
                  on_limit: approve
      responses:
        "200":
          description: Budget caps updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  agent_id:
                    type: string
                  caps:
                    type: object
                  settings:
                    type: object
                    properties:
                      on_limit:
                        type: string
                      single_tx_limit_cents:
                        type: integer
                        nullable: true
  /v1/budget/history:
    get:
      tags: [budget]
      summary: Get spend history
      description: Returns recent spend ledger entries (newest first). Cursor-paginated.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: before
          in: query
          description: ISO timestamp cursor for pagination
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Spend history
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id:
                    type: string
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        amount_usdc:
                          type: integer
                        category:
                          type: string
                        description:
                          type: string
                        reference_id:
                          type: string
                          nullable: true
                  summary:
                    type: object
                    properties:
                      count:
                        type: integer
                      total_usdc:
                        type: integer
                      by_category:
                        type: object
                  pagination:
                    type: object
                    properties:
                      limit:
                        type: integer
                      has_more:
                        type: boolean
                      next_before:
                        type: string
                        nullable: true

  # ── Charge & Approvals ─────────────────────────────────────────────────────
...